+++ mlflow/sklearn.py
@@ -185,17 +185,20 @@ def log_model(sk_model, artifact_path, conda_env=None,
                                   future release without warning. If given, create a model
                                   version under ``registered_model_name``, also creating a
                                   registered model if one with the given name does not exist.
-    :param signature: Note:: Experimental: This argument may change or be removed in a
-                      future release without warning. Model signature describes model input
-                      and output schema. The model signature can be inferred from datasets
-                      representing valid model input (e.g. the training dataset) and valid
-                      model output (e.g. model predictions generated on the training dataset).
-                      For example, you can obtain model signature as follows:
-                      ```
-                      from mlflow.models.signature import infer_signature
-                      train = df.drop_column("target_label")
-                      signature = infer_signature(train, model.predict(train))
-                      ```
+
+   :param signature: Note:: Experimental: This argument may change or be removed in a
+                     future release without warning. Model signature describes model input
+                     and output schema. The model signature can be inferred from datasets
+                     representing valid model input (e.g. the training dataset) and valid
+                     model output (e.g. model predictions generated on the training dataset).
+                     For example, you can obtain model signature as follows:
+
+                     .. code-block:: python
+
+                       from mlflow.models.signature import infer_signature
+                       train = df.drop_column("target_label")
+                       signature = infer_signature(train, model.predict(train))
+
 
     :param input_example: Note:: Experimental: This argument may change or be removed in a
                           future release without warning. Input example provides one or several
